Add support for multiple tags in mass import, whitespace to be included before each tag to be considered valid
