+++ src/main/java/seedu/address/logic/parser/exceptions/FileParseException.java
@@ -1,7 +1,5 @@
 package seedu.address.logic.parser.exceptions;
 
-import seedu.address.commons.exceptions.IllegalValueException;
-
 /**
  * Represents a parse error encountered by a file parser.
  */
